+++ src/down_search_item.d.ts
@@ -2,4 +2,4 @@ import type { search_params_I } from './_types'
 /**
  * Returns a `down_search_item` function, which sets `idx` to the next value
  */
-export declare function down_search_item_<I extends unknown = unknown>({ a$, idx$, }:search_params_I<I>):()=>void;
+export declare function down_search_item<I extends unknown = unknown>({ a$, idx$, }:search_params_I<I>):void;