+++ src/index.d.ts
@@ -1,4 +1,4 @@
 export * from './_types'
-export * from './down_search_item_'
+export * from './down_search_item'
 export * from './search_store$_'
-export * from './up_search_item_'
+export * from './up_search_item'