+++ src/index.js
@@ -1,3 +1,3 @@
-export * from './down_search_item_.js'
+export * from './down_search_item.js'
 export * from './search_store$_/index.js'
-export * from './up_search_item_.js'
+export * from './up_search_item.js'